:truck: [templates] Move render adapter from application.cookiecutter
